requirements.txt
# Core dependencies
requests>=2.31.0
python-dotenv>=1.0.0
schedule>=1.2.0
sqlalchemy>=2.0.0

# AI & Content Generation
groq>=0.4.0
openai>=1.0.0

# Trend Detection
praw>=7.7.1  # Reddit API
pytrends>=4.9.2  # Google Trends
tweepy>=4.14.0  # Twitter/X API
beautifulsoup4>=4.12.0
selenium>=4.15.0

# Text-to-Speech
TTS>=0.22.0  # Coqui TTS
edge-tts>=6.1.0
gTTS>=2.5.0

# Video & Image Processing
moviepy>=1.0.3
Pillow>=10.1.0
opencv-python>=4.8.1
imageio>=2.33.0
imageio-ffmpeg>=0.4.9

# Social Media APIs
google-api-python-client>=2.108.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
TikTokApi>=6.0.0
instagrapi>=2.0.0
facebook-sdk>=3.1.0

# Utilities
pytz>=2023.3
colorama>=0.4.6
tqdm>=4.66.1

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.1
pytest-asyncio>=0.21.0
